removed dtype casting

+++ conllu_datasets.py
@@ -107,16 +107,9 @@ class ConlluMapDataset(Dataset):
                     uposes.append(tok["upos"])
                     lemma_rules.append(l_rule)
                     ufeats.append(tok["feats"])
-                uposes = tensor(
-                    self.upos_encoder.transform(uposes), dtype=torch.float16
-                )
-                lemma_rules = tensor(
-                    self.lemma_encoder.transform(lemma_rules),
-                    dtype=torch.float16,
-                )
-                ufeats = tensor(
-                    self.ufeats_encoder.transform(ufeats), dtype=torch.float16
-                )
+                uposes = tensor(self.upos_encoder.transform(uposes))
+                lemma_rules = tensor(self.lemma_encoder.transform(lemma_rules))
+                ufeats = tensor(self.ufeats_encoder.transform(ufeats))
                 data.append((sentence, tokens, uposes, lemma_rules, ufeats))
         return data
